# [PERF] website_sale: Speedup product attribute query