### Description:

The default query to fetch attributes for searched products needs to join on the `product_tmpl_ids` Many2Many field. This can be slow, as it requires a join on an intermediate table. This change avoids that by directly querying the attribute lines and grouping by `attribute_id` to remove duplicates.

### Benchmark (in 18.0):

| N° of products | Before | After |
|----------------|--------|-------|
|           3708 |  400ms | 200ms |
|         128539 |  4 min | 800ms |
|         484798 |  6 min |    3s |

### Reference:

opw-4937361