@robodoo 2025-08-06
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/222064.png)](https://mergebot.odoo.com/odoo/odoo/pull/222064)