@robodoo 2025-08-07
@auma-odoo retry makes no sense when the PR is not in error.